EXPLOSIONS!!!
I've added a script to handle explosions. There's a button to press to
"prime" the explosion (since I didn't want to diddle with the manager
considering this is the only file I want to upload since my project
tweaked out). Drop this on the projectile and it should explode once
upon collision with an object with health. Hit the "Prime" button to
re-set it to explode again.