More Title Edits
Options screen now has Volume "Knobs"
Cube fall pattern tightened up, should prevent from cubes spawning
outside of camera view angle.

Somewhere out there in the vast nothingness of space. Somewhere far
away in space and time. Staring upwards at the gleaming stars in the
obsidian sky. A maroon on a small island. In an endless sea. Confined
to a tiny speck of sand, unable to escape. But tonight… On this small
planet.. On Earth. We're going to rock civilization.

--Thomas